Rename libretro-311fec1.h to libretro.h
