+++ rust-libretro-sys/libretro.h
